Rename services/random.ts to random-service.ts
We should make all Service API providers use the -service.ts name for clarity; this is additionally confusing because we also have runtime/random.ts
