+++ shells/configuration/whitelisted.js
@@ -28,4 +28,4 @@ import '../lib/database/firebase-upload.js';
 // TODO(sjmiles): TensorFlowJs (tfjs, also part of ml5) uses `new Function()` which requires `unsafe-eval` csp
 import '../services/textclassifier-service.js';
 import '../lib/services/tf.js';
-import '../../build/services/random.js';
+import '../../build/services/random-service.js';
